Didn't work on AVR's
Changed sizes of variables

+++ MotoPanel.h
@@ -19,13 +19,13 @@ public:
     explicit MotoPanel(Adafruit_GFX& display);
 
     // initialize beginning state
-    void begin(int rpm_range, int start_mileage);
+    void begin(int rpm_range, unsigned long start_mileage);
 
     // send values to panel to draw
     void setSpeed(int spd);
-    void setMileage(int mileage);
-    void setTrip1Mileage(int mileage);
-    void setTrip2Mileage(int mileage);
+    void setMileage(unsigned long mileage);
+    void setTrip1Mileage(unsigned long mileage);
+    void setTrip2Mileage(unsigned long mileage);
     void setRPM(int rpm);
     void setVoltage(float v);
     
@@ -41,7 +41,7 @@ public:
     
 private:
     void drawSpeed();
-    void drawMileageCore(int mileage, int len);
+    void drawMileageCore(unsigned long mileage, int len);
     void drawMileage();
     void drawTrip1Mileage();
     void drawTrip2Mileage();
@@ -51,9 +51,9 @@ private:
     Adafruit_GFX& _display;
     Mode _mode;
     int _last_speed;
-    int _last_mileage;
-    int _trip1_mileage;
-    int _trip2_mileage;
+    unsigned long _last_mileage;
+    unsigned long _trip1_mileage;
+    unsigned long _trip2_mileage;
     int _last_rpm;
     float _last_volt;
     int _rpm_range;
